@@ -31,8 +31,8 @@
     "url": "https://github.com/angular/angular-cli.git"
   },
   "engines": {
-    "node": ">= 6.9.0",
-    "npm": ">= 3.0.0"
+    "node": ">= 8.9.0",
+    "npm": ">= 5.5.1"
   },
   "author": "Angular Authors",
   "license": "MIT",
@@ -46,6 +46,7 @@
     "@angular-devkit/schematics": "0.3.1",
     "@schematics/angular": "0.3.1",
     "@schematics/package-update": "0.3.1",
+    "ajv": "^6.1.1",
     "autoprefixer": "^7.2.3",
     "cache-loader": "^1.2.0",
     "chalk": "~2.2.0",
@@ -54,9 +55,7 @@
     "common-tags": "^1.3.1",
     "copy-webpack-plugin": "~4.4.1",
     "core-object": "^3.1.0",
-    "denodeify": "^1.2.1",
     "ember-cli-string-utils": "^1.0.0",
-    "enhanced-resolve": "^3.4.1",
     "extract-text-webpack-plugin": "^3.0.2",
     "file-loader": "^1.1.5",
     "fs-extra": "^4.0.0",
@@ -110,7 +109,6 @@
     "@angular/service-worker": "^5.2.0",
     "@types/common-tags": "^1.2.4",
     "@types/copy-webpack-plugin": "^4.0.0",
-    "@types/denodeify": "^1.2.30",
     "@types/express": "^4.0.32",
     "@types/fs-extra": "^4.0.0",
     "@types/glob": "^5.0.29",
@@ -119,7 +117,7 @@
     "@types/lodash": "~4.14.50",
     "@types/minimist": "^1.2.0",
     "@types/mock-fs": "^3.6.30",
-    "@types/node": "^6.0.84",
+    "@types/node": "^8.9.1",
     "@types/request": "~2.0.0",
     "@types/semver": "^5.3.30",
     "@types/source-map": "0.5.2",
