@@ -17,8 +17,8 @@
     "url": "https://github.com/angular/angular-cli.git"
   },
   "engines": {
-    "node": ">= 6.9.0",
-    "npm": ">= 3.0.0"
+    "node": ">= 8.9.0",
+    "npm": ">= 5.5.1"
   },
   "author": "Angular Authors",
   "license": "MIT",
@@ -34,6 +34,7 @@
     "@ngtools/webpack": "1.10.0-rc.0",
     "@schematics/angular": "0.3.1",
     "@schematics/package-update": "0.3.1",
+    "ajv": "^6.1.1",
     "autoprefixer": "^7.2.3",
     "cache-loader": "^1.2.0",
     "chalk": "~2.2.0",
@@ -42,7 +43,6 @@
     "common-tags": "^1.3.1",
     "copy-webpack-plugin": "~4.4.1",
     "core-object": "^3.1.0",
-    "denodeify": "^1.2.1",
     "ember-cli-string-utils": "^1.0.0",
     "extract-text-webpack-plugin": "^3.0.2",
     "file-loader": "^1.1.5",
