@@ -22,8 +22,8 @@
   },
   "homepage": "https://github.com/angular/angular-cli/tree/master/packages/@ngtools/json-schema",
   "engines": {
-    "node": ">= 4.1.0",
-    "npm": ">= 3.0.0"
+    "node": ">= 8.9.0",
+    "npm": ">= 5.5.1"
   },
   "dependencies": {
   },
