@@ -22,8 +22,8 @@
   },
   "homepage": "https://github.com/angular/angular-cli/tree/master/packages/@ngtools/logger",
   "engines": {
-    "node": ">= 6.9.0",
-    "npm": ">= 3.0.0"
+    "node": ">= 8.9.0",
+    "npm": ">= 5.5.1"
   },
   "dependencies": {
     "rxjs": "^5.0.1"
