@@ -21,14 +21,12 @@
   },
   "homepage": "https://github.com/angular/angular-cli/tree/master/packages/@ngtools/webpack",
   "engines": {
-    "node": ">= 6.9.0",
-    "npm": ">= 3.0.0"
+    "node": ">= 8.9.0",
+    "npm": ">= 5.5.1"
   },
   "dependencies": {
     "tree-kill": "^1.0.0",
     "chalk": "~2.2.0",
-    "loader-utils": "^1.0.2",
-    "enhanced-resolve": "^3.1.0",
     "magic-string": "^0.22.3",
     "semver": "^5.3.0",
     "source-map": "^0.5.6",
